requirements.txt
bcolz==1.2.1
boto3==1.9.134
botocore==1.12.134
Django==2.1
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-storages==1.7.1
docutils==0.14
jmespath==0.9.4
numpy==1.16.2
opencv-python==4.1.0.25
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
urllib3==1.24.2
